refactor(balance): Externalize getSwitchesState to helpers
