+++ src/ducks/balance/Balance.jsx
@@ -20,36 +20,7 @@ import History from './History'
 import styles from './Balance.styl'
 import BalanceTables from './BalanceTables'
 import BalancePanels from './BalancePanels'
-
-function getSwitchesState(groups, currentSwitchesState) {
-  const switchesState = groups.reduce((acc, group) => {
-    const groupChecked = get(
-      currentSwitchesState,
-      `[${group._id}].checked`,
-      true
-    )
-
-    acc[group._id] = {
-      checked: groupChecked,
-      accounts: get(group, 'accounts.data', []).reduce((acc2, account) => {
-        acc2[account._id] = {
-          checked: get(
-            currentSwitchesState,
-            `[${group._id}].accounts[${account.id}].checked`,
-            true
-          ),
-          disabled: !groupChecked
-        }
-
-        return acc2
-      }, {})
-    }
-
-    return acc
-  }, {})
-
-  return switchesState
-}
+import { getSwitchesState } from './helpers'
 
 class Balance extends PureComponent {
   state = {
