+++ src/ducks/balance/helpers.js
@@ -150,3 +150,33 @@ export const getGroupBalance = group => {
 
   return sumBy(accounts, account => get(account, 'balance', 0))
 }
+
+export const getSwitchesState = (groups, currentSwitchesState) => {
+  const switchesState = groups.reduce((acc, group) => {
+    const groupChecked = get(
+      currentSwitchesState,
+      `[${group._id}].checked`,
+      true
+    )
+
+    acc[group._id] = {
+      checked: groupChecked,
+      accounts: get(group, 'accounts.data', []).reduce((acc2, account) => {
+        acc2[account._id] = {
+          checked: get(
+            currentSwitchesState,
+            `[${group._id}].accounts[${account.id}].checked`,
+            true
+          ),
+          disabled: !groupChecked
+        }
+
+        return acc2
+      }, {})
+    }
+
+    return acc
+  }, {})
+
+  return switchesState
+}
